moved bin cli tests into `tests/bin`
